Uploaded Files directly into GitHub
This is to create a first set of files on GitHub. It simulates the initial presence of a clients website on GitHub in your clients repo. The exercise to follow is to clone this repo to your personal computer in order to simulate your first day at work preparing to work on on your clients website at your new job.